feat(core): remove node hashing implementation
