--- packages/nx/src/hasher/impl/native-file-hasher.ts
@@ -1,42 +1,0 @@
-import { FileHasher, FileHasherBase } from './file-hasher-base';
-import { performance } from 'perf_hooks';
-import { workspaceRoot } from '../../utils/workspace-root';
-
-export class NativeFileHasher extends FileHasherBase implements FileHasher {
-  static available() {
-    try {
-      require('../../native');
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  async init(): Promise<void> {
-    performance.mark('init hashing:start');
-    // Import as needed. There is also an issue running unit tests in Nx repo if this is a top-level import.
-    const { hashFiles } = require('../../native');
-    this.clear();
-    const filesObject = hashFiles(workspaceRoot);
-    this.fileHashes = new Map(Object.entries(filesObject));
-
-    performance.mark('init hashing:end');
-    performance.measure(
-      'init hashing',
-      'init hashing:start',
-      'init hashing:end'
-    );
-  }
-
-  hashFile(path: string): string {
-    // Import as needed. There is also an issue running unit tests in Nx repo if this is a top-level import.
-    const { hashFile } = require('../../native');
-    return hashFile(path).hash;
-  }
-}
-
-export function nativeHashArray(input: string[]): string {
-  // Import as needed. There is also an issue running unit tests in Nx repo if this is a top-level import.
-  const { hashArray } = require('../../native');
-  return hashArray(input);
-}
