+++ packages/nx/src/plugins/js/lock-file/npm-parser.ts
@@ -8,7 +8,7 @@ import {
   ProjectGraph,
   ProjectGraphExternalNode,
 } from '../../../config/project-graph';
-import { fileHasher, hashArray } from '../../../hasher/impl';
+import { fileHasher, hashArray } from '../../../hasher/file-hasher';
 
 /**
  * NPM
