+++ packages/nx/src/plugins/js/lock-file/yarn-parser.ts
@@ -8,7 +8,7 @@ import {
   ProjectGraph,
   ProjectGraphExternalNode,
 } from '../../../config/project-graph';
-import { hashArray } from '../../../hasher/impl';
+import { hashArray } from '../../../hasher/file-hasher';
 import { sortObjectByKeys } from '../../../utils/object-sort';
 
 /**
