+++ packages/nx/src/utils/all-file-data.ts
@@ -1,6 +1,6 @@
 import { FileData } from '../config/project-graph';
 import { daemonClient } from '../daemon/client/client';
-import { fileHasher } from '../hasher/impl';
+import { fileHasher } from '../hasher/file-hasher';
 
 export function allFileData(): Promise<FileData[]> {
   if (daemonClient.enabled()) {
